+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Derive target filter when an entity with graph link updates and the target has a dynamic filter.
+
 ### [v6.177](https://github.com/replicant4j/replicant/tree/v6.177) (2026-03-06) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.176...v6.177)
 
 Changes in this release: